requirements.txt
autopep8==1.6.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
flake8==4.0.1
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-bigquery==2.32.0
google-cloud-core==2.2.2
google-cloud-secret-manager==2.15.1
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.2.1
googleapis-common-protos==1.58.0
grpc-google-iam-v1==0.12.6
grpcio==1.51.1
grpcio-status==1.43.0
httplib2==0.20.4
idna==3.3
mccabe==0.6.1
numpy==1.22.2
oauthlib==3.2.0
packaging==21.3
pandas==1.4.1
proto-plus==1.22.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.8
